Tinkering with c++/python api.
Former-commit-id: b0ba792de13c4d659d0ed7d77c903221b3033770
Former-commit-id: 6e8b76ec15b1554e8cb9ddf546a82bd51e5ca4ac

+++ python_port/dot_shooting.h
@@ -1,13 +1,13 @@
-void init(int x, int y, int random_bg,int max_moves,float living_reward,float miss_penalty, float hit_reward, int ammo);
+void init(int x, int y, int random_bg,int max_moves,double living_reward,double miss_penalty, double hit_reward, int ammo);
 
 int * get_state_format();
 int get_action_format();
 
 void new_episode();
-float make_action( int const* action);
+double make_action( int const* action);
 int is_finished();
-float get_summary_reward();
+double get_summary_reward();
 float * get_image_state();
 float * get_misc_state();
 
-float average_best_result();
+double average_best_result();
